search_module/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pdfplumber
scikit-learn
numpy
pyyaml 